examples/langchain/instrumentation-comparison/braintrust-sdk/pyproject.toml
[project]
name = "langchain-weather-braintrust"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "langchain-aws>=0.2.0",
    "langchain-core>=0.3.0",
    "braintrust[otel]>=0.3.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
